dashboard/requirements.txt
Flask==2.3.3
oracledb==1.4.2
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0 